requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pypdf==3.17.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
tiktoken==0.5.2
alembic==1.13.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
